move number range to data components
