+++ app/src/frontend/building/data-components/number-range-data-entry.css
